+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented here. The format is based
 on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project
 adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Features
+
+- optimised server error handling
+
+### Chore
+
+- added travis and tests
+
 ## 0.1.0 (2020-04-12)
 
 ### Features